feat: restructure forgot password flow by moving components to features directory and adding new screens
